pyproject.toml
[project]
name = "chat-webapp"
version = "1.0.0"
description = ""
authors = [
    { name = "Alexis Houssemand", email = "alexis.houssemand@gmail.com" },
]
license = { text = "" }
readme = "README.md"
requires-python = ">=3.12"


dependencies = [
    "django >=5.1.6,<6.0.0",
    "djangorestframework >=3.15.2,<4.0.0",
    "redis >=5.2.1,<6.0.0",
    "psycopg2-binary >=2.9.10,<3.0.0",
    "watchdog >=6.0.0,<7.0.0",
    "requests >=2.32.3,<3.0.0",
    "channels >=4.2.0,<5.0.0",
    "channels-redis >=4.2.1,<5.0.0",
    "pillow >=11.1.0,<12.0.0",
    "gunicorn >=23.0.0,<24.0.0",
    "uvicorn[standard] >=0.34.0,<0.35.0",
    "xlsxwriter >=3.2.2,<4.0.0",
    "boto3 >=1.36.14,<2.0.0",
    "django-cors-headers >=4.6.0,<5.0.0",
    "tsidpy >=1.1.5,<2.0.0",
    "djangorestframework-simplejwt >=5.4.0,<6.0.0",
    "argh >=0.31.3,<0.32.0",
]


[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.4,<9.0.0"
pytest-django = ">=4.9.0,<5.0.0"
pytest-factoryboy = ">=2.7.0,<3.0.0"
pytest-cov = ">=6.0.0,<7.0.0"
factory-boy = ">=3.3.3,<4.0.0"
black = ">=24.0.0,<25.0.0"
mypy = ">=1.10.0,<2.0.0"
isort = ">=5.12.0,<6.0.0"
flake8 = ">=6.1.0,<7.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
